Make /en etc routable without client-side redirect
