+++ examples/app-dir-i18n-routing/app/[lang]/components/counter.tsx
