+++ examples/app-dir-i18n-routing/app/[lang]/components/locale-switcher.tsx
@@ -2,7 +2,7 @@
 
 import { usePathname } from 'next/navigation'
 import Link from 'next/link'
-import { i18n } from '../../../../../i18n-config'
+import { i18n } from '../../../i18n-config'
 
 export default function LocaleSwitcher() {
   const pathName = usePathname()
