+++ examples/app-dir-i18n-routing/app/[lang]/head.tsx
