--- examples/app-dir-i18n-routing/app/[lang]/some/custom/path/page.tsx
@@ -1,24 +1,0 @@
-import { getDictionary } from '../../../../../get-dictionary'
-import { Locale } from '../../../../../i18n-config'
-import Counter from './counter'
-import LocaleSwitcher from './locale-switcher'
-
-export default async function IndexPage({
-  params: { lang },
-}: {
-  params: { lang: Locale }
-}) {
-  const dictionary = await getDictionary(lang)
-
-  return (
-    <div>
-      <LocaleSwitcher />
-      <p>Current locale: {lang}</p>
-      <p>
-        This text is rendered on the server:{' '}
-        {dictionary['server-component'].welcome}
-      </p>
-      <Counter dictionary={dictionary.counter} />
-    </div>
-  )
-}
